colisions fixed

+++ include/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: saroca-f <saroca-f@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/19 18:26:26 by schamizo          #+#    #+#             */
-/*   Updated: 2024/09/05 10:09:52 by saroca-f         ###   ########.fr       */
+/*   Updated: 2024/09/05 16:25:47 by saroca-f         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -198,7 +198,7 @@ void	print_border_error(int line, int flag);
 
 //game
 
-bool	check_cell_move(mlx_image_t *image, t_data *data, int move_x, int move_y);
+bool	check_cell_move(mlx_image_t *image, t_data *data, double move_x, double move_y);
 void	ft_game(t_data *data);
 void	print_ray(t_data *data, t_player *player);
